requirements.txt
# Core dependencies
pdfplumber>=0.10.3
pandas>=2.0.0
python-dateutil>=2.8.2

# AI/ML
requests>=2.31.0
ollama>=0.1.0

# Data validation
jsonschema>=4.20.0

# CLI
click>=8.1.7
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Visualization (optional)
plotly>=5.18.0
matplotlib>=3.8.0
